eth_analyzer/requirements.txt
web3
eth-abi
eth-utils
python-dotenv

